Using "extra flash" for config storage
The STM32F103C8T6 is advertised as a chip with 64k of flash memory but it have 128k. flash_storage.h exploits this extra memory to save the config on a segment that will not be rewritten when updating the program. When the chip will be flashed with a new "64k" program, the config will be safe on the extra memory and won't be overwritten.